// src/data/form.pb.ts
export type AcceptStatus = "WAITING" | "REJECTED" | "ACCEPTED";
export const AcceptStatus_WAITING: AcceptStatus = "WAITING";
export const AcceptStatus_REJECTED: AcceptStatus = "REJECTED";
export const AcceptStatus_ACCEPTED: AcceptStatus = "ACCEPTED";

export const ALL_AcceptStatus_VALUES: AcceptStatus[] = [
  AcceptStatus_WAITING,
  AcceptStatus_REJECTED,
  AcceptStatus_ACCEPTED,
];

export type ListingAppOrderBy =
  | "ADDED_NEW_TO_OLD"
  | "ADDED_OLD_TO_NEW"
  | "MINTED_NEW_TO_OLD"
  | "MINTED_OLD_TO_NEW"
  | "TOTAL_ITEMS_LOW_TO_HIGH"
  | "TOTAL_ITEMS_HIGH_TO_LOW";
// ...
export interface ListingApp {
  /**
unique id
*/
  id?: string;
  /**
public key of application submitter
TODO: only admin can update
*/
  ownerPubkey?: string;
  /**
whether the project is reached by us or not
TODO: only admin can update
*/
  reached?: boolean;
  /**
title of collection
*/
  name?: string;
  /**
email address for contact
*/
  email?: string;
  /**
project twitter url
*/
  twitter?: string;
  /**
discord server invite link
*/
  discordInvite?: string;
  /**
discord id (nickname) for contact
*/
  discordId?: string;
  /**
short description of the project (for the collection page)
*/
  description?: string;
  /**
creator or Artist Name
*/
  artist?: string;
  /**
project website, instagram etc
*/
  otherLinks?: string[];
  /**
is collection already minted (false if partially minted)
*/
  minted?: boolean;
  /**
candy machine ids
*/
  candyIds?: string[];
  /**
the image that is shown on collection profile page (500x500)
*/
  pfp?: string;
  /**
"upcoming launches" (400x300)
*/
  featuredImage?: string;
  /**
banner image (1600x600)
*/
  banner?: string;
  /**
number of items in collection (existing or expected)
*/
  totalItems?: number;
  /**
anything else
*/
  other?: string;
  /**
whether the application is accepted or not
*/
  accepted?: boolean;
  /**
submission date
*/
  addedAt?: google_protobuf_google_protobuf_timestamp.Timestamp;
  /**
accept date, null if not accepted yet or rejected
*/
  acceptedAt?: google_protobuf_google_protobuf_timestamp.Timestamp;
  /**
mint date, null if not minted or partially minted
*/
  mintTime?: google_protobuf_google_protobuf_timestamp.Timestamp;
  /**
whether featured on the platform
*/
  featured?: boolean;
  /**
start of feature
*/
  featureFrom?: google_protobuf_google_protobuf_timestamp.Timestamp;
  /**
end of feature
*/
  featureUntil?: google_protobuf_google_protobuf_timestamp.Timestamp;
  acceptStatus?: AcceptStatus;
}
// ...
export interface ListingAppMatch {
  ownerPubkey?: string;
  name?: string;
  reached?: boolean;
  /**
filter by multiple status
*/
  acceptStatus?: AcceptStatus[];
  minted?: boolean;
  /**
select from before or after + added_at or mint_time
not included in match_fields, time.time == nil if empty
*/
  time?: ListingAppSelectTime;
  /**
backwards compat. reasons
*/
  accepted?: boolean;
}

export const ListingAppMatch_ownerPubkey = "owner_pubkey";
export const ListingAppMatch_name = "name";
export const ListingAppMatch_reached = "reached";
export const ListingAppMatch_acceptStatus = "accept_status";
export const ListingAppMatch_minted = "minted";
export const ListingAppMatch_time = "time";
export const ListingAppMatch_accepted = "accepted";
/**
select via added_at or mint_time
*/
export interface ListingAppSelectTime {
  type?: ListingAppTimeType;
  time?: google_protobuf_google_protobuf_timestamp.Timestamp;
}

export const ListingAppSelectTime_type = "type";
export const ListingAppSelectTime_time = "time";
export interface FilterListingAppReq {
  orderBy?: ListingAppOrderBy;
  matchFields?: google_protobuf_google_protobuf_field_mask.FieldMask;
  matchValues?: ListingAppMatch;
  limit?: number;
  offset?: number;
}